Moved tenant related documents to tenant folder
