+++ openframe-data-mongo/src/main/java/com/openframe/data/document/tenant/SSOPerTenantConfig.java
@@ -1,5 +1,6 @@
-package com.openframe.data.document.sso;
+package com.openframe.data.document.tenant;
 
+import com.openframe.data.document.sso.SSOConfig;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
