+++ openframe-data-mongo/src/main/java/com/openframe/data/repository/tenant/TenantClusterRegistrationRepository.java
@@ -1,0 +1,9 @@
+package com.openframe.data.repository.tenant;
+
+import com.openframe.data.document.tenant.TenantClusterRegistration;
+import org.springframework.data.mongodb.repository.MongoRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface TenantClusterRegistrationRepository extends MongoRepository<TenantClusterRegistration, String> {
+}
