+++ openframe-data-mongo/src/main/java/com/openframe/data/repository/tenant/TenantRepository.java
@@ -1,6 +1,6 @@
-package com.openframe.data.repository.auth;
+package com.openframe.data.repository.tenant;
 
-import com.openframe.data.document.auth.Tenant;
+import com.openframe.data.document.tenant.Tenant;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
